refactor (api): move pino instance to own file
