+++ server/src/index.ts
@@ -1,3 +1,4 @@
+import { ApolloServerPluginLandingPageGraphQLPlayground } from "apollo-server-core";
 import { ApolloServer } from "apollo-server-express";
 import connectRedis from "connect-redis";
 import cors from "cors";
@@ -11,40 +12,14 @@ import config from "./config";
 import typeormConfig from "./ormconfig";
 import { resolvers } from "./resolvers";
 import { authChecker } from "./utils/authChecker";
-import { ApolloServerPluginLandingPageGraphQLPlayground } from "apollo-server-core";
-import pinoHttp from "pino-http";
-import childProcess from "child_process";
-import stream from "stream";
-
-const cwd = process.cwd();
-const { env } = process;
-const logThrough = new stream.PassThrough();
-const logsPath = `${cwd}/logs`;
-
-const child = childProcess.spawn(
-  process.execPath,
-  [
-    require.resolve("pino-tee"),
-    "warn",
-    `${logsPath}/warn.log`,
-    "error",
-    `${logsPath}/error.log`,
-    "fatal",
-    `${logsPath}/fatal.log`,
-    "debug",
-    `${logsPath}/debug.log`,
-  ],
-  { cwd, env }
-);
-
-logThrough.pipe(child.stdin);
+import { pinoInstance } from "./utils/logger";
 
 const main = async () => {
   await createConnection(typeormConfig);
   // await conn.runMigrations();
   const app = express();
-  const pino = pinoHttp({ name: "openco2roadmap" }, logThrough);
-  app.use(pino);
+
+  app.use(pinoInstance);
   app.use(express.json());
 
   const RedisStore = connectRedis(session);
